refactor: split builtins.js (2640 lines) into tool-domain modules
Single-file dispatch with 47 _exec* handlers in one 2640-line switch
made it hard to add or audit individual tools. Split by domain into a
builtins/ subdirectory:

- _shared.js: cross-domain helpers (incl. 430-line _executePageAction
  injection script reused by dom and tabs)
- tabs.js, dom.js, groups.js, windows.js, history.js, downloads.js,
  stash.js, macros.js, schedule.js, evalAndPlayground.js, misc.js

The entry builtins.js (181 lines) holds only MCP routing, timeout
wrapping, and a BUILTIN_TOOL_HANDLERS lookup table replacing the
50-case switch — adding a tool now only touches its domain file and
one dictionary entry. Public exports (executeTool,
getBuiltinToolTimeoutSeconds, captureFullPageScreenshotToTab,
openHelloWorldPlayground) preserved; no caller changes needed.
